Put reshaping methods in base class
