+++ darts/dataprocessing/transformers/base_data_transformer.py
@@ -3,6 +3,8 @@ Base Data Transformer
 ---------------------
 """
 from typing import Sequence, Union, Iterator, Tuple, List
+import numpy as np
+
 from darts.logging import raise_if_not
 from darts.utils import _parallel_apply, _build_tqdm_iterator
 from darts import TimeSeries
@@ -174,6 +176,26 @@ class BaseDataTransformer(ABC):
                                            self._n_jobs, args, kwargs)
 
         return transformed_data[0] if isinstance(series, TimeSeries) else transformed_data
+
+    @staticmethod
+    def _reshape_in(series: TimeSeries) -> np.ndarray:
+        """ Reshapes the series' values to be fed in input to a transformer.
+
+            The output is a 2-D matrix where each column corresponds to a component (dimension)
+            of the series, and the columns' values are the flattened values over all samples
+        """
+        vals = series.all_values(copy=False)
+        return np.stack([vals[:, i, :].reshape(-1) for i in range(series.width)], axis=1)
+
+    @staticmethod
+    def _reshape_out(vals: np.ndarray, series_width: int, series_n_samples: int) -> np.ndarray:
+        """ Reshapes the 2-D matrix coming out of a transformer into a 3-D matrix
+            suitable to build a TimeSeries.
+
+            The output is a 3-D matrix, built by taking each column of the 2-D matrix (the flattened components)
+            and reshaping them to (len(series), n_samples), then stacking them on 2nd axis.
+        """
+        return np.stack([vals[:, i].reshape(-1, series_n_samples) for i in range(series_width)], axis=1)
 
     @property
     def name(self):
